// src/FE/src/api/generated/models/customerSearchViewModel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CustomerSearchViewModel {
  id: string;
  /** @nullable */
  customerNumber: string | null;
  name: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  address: string | null;
  /** @nullable */
  zipCode: string | null;
  /** @nullable */
  city: string | null;
  /** @nullable */
  country: string | null;
  /** @nullable */
  contactPerson: string | null;
  isFavorite: boolean;
}