// src/FE/src/api/generated/models/customerSnapshotData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CustomerSnapshotData {
  /** @nullable */
  name: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  address: string | null;
  /** @nullable */
  contactPerson: string | null;
}