// src/FE/src/api/generated/models/userViewModel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface UserViewModel {
  id: string;
  organizationId: string;
  email: string;
  displayName: string;
  phone: string;
  role: string;
  roleDisplayName: string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  hoursThisWeek: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  hoursThisMonth: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  hoursBiweekly: number | string | null;
}